Improve and simplify build_point_cloud()
This change applies changes suggested by Joel. It also gets rid of
glam library dependency for the base and the composer crates, solely
relying on nalgebra.

+++ composer/src/optimize_scan_geometry.rs
@@ -222,7 +222,7 @@ impl<'a> ArgminOp for ScanOpt<'a> {
             }
         }
 
-        Ok(sum / num as f32)
+        Ok(sum as f32 / num as f32)
     }
 
     fn gradient(&self, p: &Self::Param) -> StdResult<Self::Param, ArgminError> {
